collapse extensions into single class
